Excercise  Challenge
This is the first real mini project i did within the course i'm currently taking of python 3, took quite a while of time to figure it out. The syntax of the comments i based it of the solution provided by the course, and the enhances to the code, my original seemed like potato.